........
r11503 | tilghman | 2006-03-01 11:41:52 -0600 (Wed, 01 Mar 2006) | 2 lines

Bug 6615 - Fix 64bit conversion errors by using a long int

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@11504 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/cdr.h
@@ -74,13 +74,13 @@ struct ast_cdr {
 	
 	struct timeval end;
 	/*! Total time in system, in seconds */
-	int duration;				
+	long int duration;				
 	/*! Total time call is up, in seconds */
-	int billsec;				
+	long int billsec;				
 	/*! What happened to the call */
-	int disposition;			
+	long int disposition;			
 	/*! What flags to use */
-	int amaflags;				
+	long int amaflags;				
 	/*! What account number to use */
 	char accountcode[AST_MAX_ACCOUNT_CODE];			
 	/*! flags */
